+++ pkg/R/rlassoEffects.R
@@ -102,7 +102,7 @@ rlassoEffects <- function(x, y, index = c(1:ncol(x)), method = "partialling out"
     Xt <- x[, -index[i], drop = FALSE]
     I3m <- I3[-index[i]]
     lasso.regs[[i]] <- try(col <- rlassoEffect(Xt, y, d, method = method, 
-                                               I3 = I3m, post = post, ...))
+                                               I3 = I3m, post = post, ...), silent = TRUE)
     if (class(lasso.regs[[i]]) == "try-error") {
       next
     } else {
@@ -153,6 +153,7 @@ rlassoEffect <- function(x, y, d, method = "double selection", I3 = NULL,
     x <- cbind(d, x[, I, drop = FALSE])
     reg1 <- lm(y ~ x)
     alpha <- coef(reg1)[2]
+    names(alpha) <- colnames(d)
     xi <- reg1$residuals * sqrt(n/(n - sum(I) - 1))
     if (is.null(I)) {
       reg2 <- lm(d ~ 1)
@@ -239,6 +240,53 @@ print.rlassoEffects <- function(x, digits = max(3L, getOption("digits") -
 #' @param joint logical, if \code{TRUE} joint confidence intervals are calculated.
 #' @export
 
+# confint.rlassoEffects <- function(object, parm, level = 0.95, joint = FALSE, 
+#                                   ...) {
+#   B <- 500  # number of bootstrap repitions
+#   n <- object$samplesize
+#   k <- p1 <- length(object$coefficients)
+#   cf <- coef(object)
+#   pnames <- names(cf)
+#   if (missing(parm)) 
+#     parm <- pnames else if (is.numeric(parm)) 
+#       parm <- pnames[parm]
+#   if (!joint) {
+#     a <- (1 - level)/2
+#     a <- c(a, 1 - a)
+#     # fac <- qt(a, n-k)
+#     fac <- qnorm(a)
+#     pct <- format.perc(a, 3)
+#     ci <- array(NA, dim = c(length(parm), 2L), dimnames = list(parm, 
+#                                                                pct))
+#     ses <- object$se[parm]
+#     ci[] <- cf[parm] + ses %o% fac
+#   }
+#   
+#   if (joint) {
+#     phi <- object$residuals$e * object$residuals$v
+#     m <- 1/sqrt(colMeans(phi^2))
+#     phi <- t(t(phi)/m)
+#     sigma <- sqrt(colMeans(phi^2))
+#     sim <- vector("numeric", length = B)
+#     for (i in 1:B) {
+#       xi <- rnorm(n)
+#       phi_temp <- phi * xi
+#       Nstar <- 1/sqrt(n) * colSums(phi_temp)
+#       sim[i] <- max(abs(Nstar))
+#     }
+#     a <- (1 - level)/2
+#     ab <- c(a, 1 - a)
+#     pct <- format.perc(ab, 3)
+#     ci <- array(NA, dim = c(length(parm), 2L), dimnames = list(parm, 
+#                                                                pct))
+#     hatc <- quantile(sim, probs = 1 - a)
+#     ci[, 1] <- cf[parm] - hatc * 1/sqrt(n) * sigma
+#     ci[, 2] <- cf[parm] + hatc * 1/sqrt(n) * sigma
+#   }
+#   return(ci)
+# }
+
+
 confint.rlassoEffects <- function(object, parm, level = 0.95, joint = FALSE, 
                                   ...) {
   B <- 500  # number of bootstrap repitions
@@ -262,16 +310,23 @@ confint.rlassoEffects <- function(object, parm, level = 0.95, joint = FALSE,
   }
   
   if (joint) {
-    phi <- object$residuals$e * object$residuals$v
-    m <- 1/sqrt(colMeans(phi^2))
-    phi <- t(t(phi)/m)
-    sigma <- sqrt(colMeans(phi^2))
+    e <- object$residuals$e
+    v <- object$residuals$v
+    ev <- e*v
+    Ev2 <- colMeans(v^2)
+    Ee2v2 <- colMeans(ev^2)
+    Omegahat <- matrix(NA, ncol=k, nrow=k)
+    for (j in 1:k) {
+      for (l in 1:k) {
+        Omegahat[j,l] = Omegahat[l,j] = 1/(Ev2[j]*Ev2[l]) * mean(ev[,j]*ev[,l])
+        }
+    }
+    var <- diag(Omegahat)
+    Beta <- matrix(NA, ncol=B, nrow=k)
     sim <- vector("numeric", length = B)
     for (i in 1:B) {
-      xi <- rnorm(n)
-      phi_temp <- phi * xi
-      Nstar <- 1/sqrt(n) * colSums(phi_temp)
-      sim[i] <- max(abs(Nstar))
+      beta_i <- MASS::mvrnorm(mu = rep(0,k), Sigma=Omegahat/n)
+      sim[i] <- max(abs(sqrt(n)*beta_i/var))
     }
     a <- (1 - level)/2
     ab <- c(a, 1 - a)
@@ -279,11 +334,12 @@ confint.rlassoEffects <- function(object, parm, level = 0.95, joint = FALSE,
     ci <- array(NA, dim = c(length(parm), 2L), dimnames = list(parm, 
                                                                pct))
     hatc <- quantile(sim, probs = 1 - a)
-    ci[, 1] <- cf[parm] - hatc * 1/sqrt(n) * sigma
-    ci[, 2] <- cf[parm] + hatc * 1/sqrt(n) * sigma
+    ci[, 1] <- cf[parm] - hatc * 1/sqrt(n) * sqrt(var)
+    ci[, 2] <- cf[parm] + hatc * 1/sqrt(n) * sqrt(var)
   }
   return(ci)
 }
+
 
 #' @rdname methods.rlassoEffects
 #' @export
@@ -291,11 +347,11 @@ confint.rlassoEffects <- function(object, parm, level = 0.95, joint = FALSE,
 #' @param xlab a title for the x axis
 #' @param ylab a title for the y axis
 #' @param xlim vector of length two giving lower and upper bound of x axis
-plot.rlassoEffects <- function(x, main = "", xlab = "coef", ylab = "", 
+plot.rlassoEffects <- function(x, joint=FALSE, level= 0.95, main = "", xlab = "coef", ylab = "", 
                                xlim = NULL, ...) {
   
   # generate ordered KI-matrix
-  coefmatrix <- cbind(summary(x)$coef, confint(x))[, c(1, 5, 6)]
+  coefmatrix <- cbind(summary(x)$coef, confint(x, joint = joint, level=level))[, c(1, 5, 6)]
   if (is.null(dim(coefmatrix))) {
     vec <- coefmatrix
     coefmatrix <- matrix(vec, ncol = 3)
@@ -319,12 +375,12 @@ plot.rlassoEffects <- function(x, main = "", xlab = "coef", ylab = "",
   }
   # generate points
   plotobject <- ggplot2::ggplot(coefmatrix, ggplot2::aes(y = coef, x = factor(names, 
-                                                                              levels = names))) + ggplot2::geom_point(colour = col, size = 1.75) + 
-    ggplot2::geom_hline(h = 0, colour = col, width = 0.1)
+                                                                              levels = names))) + ggplot2::geom_point(colour = col) + 
+  ggplot2::geom_hline(colour = col, aes(width = 0.1, h = 0, yintercept=0))
   
   # generate errorbars (KIs)
   plotobject <- plotobject + ggplot2::geom_errorbar(ymin = coefmatrix$lower, 
-                                                    ymax = coefmatrix$upper, colour = col, width = 0.4, size = 0.2)
+                                                    ymax = coefmatrix$uppe, colour = col)
   
   # further graphic parameter
   plotobject <- plotobject + ggplot2::ggtitle(main) + ggplot2::ylim(low, 
@@ -337,6 +393,7 @@ plot.rlassoEffects <- function(x, main = "", xlab = "coef", ylab = "",
   # layout
   plotobject <- plotobject + ggplot2::theme_bw() + ggplot2::geom_blank() + 
     ggplot2::theme(panel.grid.major.x = ggplot2::element_blank(), panel.grid.minor.x = ggplot2::element_blank())
+  plotobject <- plotobject + scale_x_discrete(labels = abbreviate)
   # plot
   plotobject
 }