+++ pkg/R/rlassoIVselectX.R
@@ -39,11 +39,11 @@ rlassoIVselectX <- function(x,d,y,z, post=TRUE, ...) {
   lasso.d.x <- rlasso(d ~ x, post=post, ...)
   Dr <- d - predict(lasso.d.x, newdata=x)
   lasso.y.x <- rlasso(y ~ x, post=post, ...)
-  Yr <- y - predict(lasso.y.x, newdata=x)
+  Yr <- y - predict(lasso.y.x)
   Zr <- matrix(NA, nrow=n, ncol=numIV)
   for (i in seq(length.out=numIV)) {
   lasso.z.x <- rlasso(z[,i] ~ x, post=post, ...)
-  Zr[,i] <- z - predict(lasso.z.x, newdata=x)
+  Zr[,i] <- z - predict(lasso.z.x)
   }
   result <- tsls(Yr,Dr,x=NULL,Zr, intercept=FALSE)
   coef <- as.vector(result$coefficient)